move command back into commands
