+++ src/commands.rs
@@ -1,16 +1,74 @@
 use crate::{
     api,
-    command::{Auth, Command, CommandKind, Handler},
     state_machine::{CharacterSet, StateMachine},
+    Error,
 };
 use indexmap::IndexMap;
 use reqwest::Client as HttpClient;
 use serenity::{model::channel::Message, prelude::Context};
 use sqlx::postgres::PgPool;
-use std::{collections::HashMap, sync::Arc};
+use std::{collections::HashMap, future::Future, pin::Pin, sync::Arc};
 use tracing::{error, info};
 
 pub const PREFIX: &str = "?";
+
+type ResultFuture<T, E> = Pin<Box<dyn Future<Output = Result<T, E>> + Send>>;
+
+pub trait AsyncFn<T>: 'static {
+    fn call(&self, args: Arc<Args>) -> ResultFuture<T, Error>;
+}
+
+impl<F, G, T> AsyncFn<T> for F
+where
+    F: Fn(Arc<Args>) -> G + 'static,
+    G: Future<Output = Result<T, Error>> + Send + 'static,
+{
+    fn call(&self, args: Arc<Args>) -> ResultFuture<T, Error> {
+        let fut = (self)(args);
+        Box::pin(async move { fut.await })
+    }
+}
+
+pub type Handler = dyn AsyncFn<()> + Send + Sync;
+pub type Auth = dyn AsyncFn<bool> + Send + Sync;
+
+pub enum CommandKind {
+    Base,
+    Protected,
+    Help,
+}
+
+pub struct Command {
+    pub kind: CommandKind,
+    pub auth: &'static Auth,
+    pub handler: &'static Handler,
+}
+
+impl Command {
+    pub fn new(handler: &'static Handler) -> Self {
+        Self {
+            kind: CommandKind::Base,
+            auth: &|_| async { Ok(true) },
+            handler,
+        }
+    }
+
+    pub fn new_with_auth(handler: &'static Handler, auth: &'static Auth) -> Self {
+        Self {
+            kind: CommandKind::Protected,
+            auth,
+            handler,
+        }
+    }
+
+    pub fn help() -> Self {
+        Self {
+            kind: CommandKind::Help,
+            auth: &|_| async { Ok(true) },
+            handler: &|_| async { Ok(()) },
+        }
+    }
+}
 
 pub struct Args {
     pub cx: Context,
